movies_admin/requirements.txt
asgiref==3.8.1
Django==4.2.11
django-split-settings==1.3.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
django-debug-toolbar
